% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{predict.hierNet.logistic}
\alias{predict.hierNet.logistic}
\title{Prediction function for hierNet.logistic.}
\usage{
\method{predict}{hierNet.logistic}(object, newx, newzz = NULL, ...)
}
\arguments{
\item{object}{The results of a call to the "hierNet.logistic" or
"hierNet.logistic.path" or function. The coefficients that are part of this
object will be used for making predictions.}

\item{newx}{The new x at which predictions should be made. Can be a vector
or a matrix (one observation per row).}

\item{newzz}{Optional matrix of products of columns of newx, computed by
compute.interactions.c}

\item{...}{additional arguments (not currently used)}
}
\value{
\item{yhat}{Matrix of predictions (probabilities), one row per
observation}
}
\description{
A function to perform prediction, using an x matrix and the output of the
"hierNet.logistic" function or "hierNet.logistic.path".
}
\examples{

set.seed(12)
x=matrix(rnorm(100*10),ncol=10)
x=scale(x,TRUE,TRUE)
y=x[,1]+2*x[,2]+ x[,1]*x[,2]+3*rnorm(100)
y=1*(y>0)
newx=matrix(rnorm(100*10),ncol=10)
fit=hierNet.logistic(x,y,lam=5)
yhat=predict(fit,newx)

fit=hierNet.logistic.path(x,y)
yhat=predict(fit,newx)

}
\references{
Bien, J., Taylor, J., Tibshirani, R., (2013) "A Lasso for
Hierarchical Interactions." Annals of Statistics. 41(3). 1111-1141.
}
\seealso{
\link{hierNet.logistic}, \link{hierNet.logistic.path}
}
\author{
Jacob Bien and Robert Tibshirani
}
